requirements.txt
flask
flask_sqlalchemy
flask_api
sqlalchemy
sqlalchemy-migrate
keras
tensorflow
pandas
webcolors==1.3